+++ app/Http/Controllers/IngredientController.php
@@ -94,13 +94,4 @@ class IngredientController extends Controller
 
         return response(null, 204);
     }
-
-    public function find(Request $request): JsonResource
-    {
-        $name = $request->get('name');
-
-        $ingredient = Ingredient::with('cocktails', 'images', 'varieties', 'parentIngredient')->whereRaw('lower(name) = ?', [strtolower($name)])->firstOrFail();
-
-        return new IngredientResource($ingredient);
-    }
 }