+++ tests/Feature/CocktailControllerTest.php
@@ -10,6 +10,7 @@ use Kami\Cocktail\Models\Glass;
 use Kami\Cocktail\Models\Cocktail;
 use Kami\Cocktail\Models\Ingredient;
 use Kami\Cocktail\Models\CocktailMethod;
+use Kami\Cocktail\Models\CocktailFavorite;
 use Kami\Cocktail\Models\CocktailIngredient;
 use Illuminate\Testing\Fluent\AssertableJson;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -27,6 +28,77 @@ class CocktailControllerTest extends TestCase
         $this->actingAs(
             User::factory()->create()
         );
+    }
+
+    public function test_cocktails_response()
+    {
+        Cocktail::factory()->count(45)->create();
+
+        $response = $this->getJson('/api/cocktails');
+
+        $response->assertStatus(200);
+        $response->assertJsonCount(15, 'data');
+        $response->assertJsonPath('meta.current_page', 1);
+        $response->assertJsonPath('meta.last_page', 3);
+        $response->assertJsonPath('meta.per_page', 15);
+        $response->assertJsonPath('meta.total', 45);
+
+        $response = $this->getJson('/api/cocktails?page=2');
+        $response->assertJsonPath('meta.current_page', 2);
+
+        $response = $this->getJson('/api/cocktails?per_page=5');
+        $response->assertJsonPath('meta.last_page', 9);
+    }
+
+    public function test_cocktails_response_filters()
+    {
+        $user = User::factory()->create();
+        Cocktail::factory()->createMany([
+            ['name' => 'Old Fashioned'],
+            ['name' => 'XXXX'],
+            ['name' => 'Test', 'user_id' => $user->id],
+            ['name' => 'public', 'public_id' => 'UUID'],
+        ]);
+        Cocktail::factory()->hasTags(1)->create(['name' => 'test 1']);
+        $cocktailFavorited = Cocktail::factory()->create();
+
+        $favorite = new CocktailFavorite();
+        $favorite->cocktail_id = $cocktailFavorited->id;
+        auth()->user()->favorites()->save($favorite);
+
+        $response = $this->getJson('/api/cocktails?filter[name]=old');
+        $response->assertJsonCount(1, 'data');
+        $response = $this->getJson('/api/cocktails?filter[name]=old,xx');
+        $response->assertJsonCount(2, 'data');
+        $response = $this->getJson('/api/cocktails?filter[tag_id]=1');
+        $response->assertJsonCount(1, 'data');
+        $response = $this->getJson('/api/cocktails?filter[user_id]=' . $user->id);
+        $response->assertJsonCount(1, 'data');
+        $response = $this->getJson('/api/cocktails?filter[on_shelf]=true');
+        $response->assertJsonCount(0, 'data');
+        $response = $this->getJson('/api/cocktails?filter[favorites]=true');
+        $response->assertJsonCount(1, 'data');
+        $response = $this->getJson('/api/cocktails?filter[is_public]=true');
+        $response->assertJsonCount(1, 'data');
+    }
+
+    public function test_cocktails_response_sorts()
+    {
+        Cocktail::factory()->createMany([
+            ['name' => 'B Cocktail'],
+            ['name' => 'A Cocktail'],
+            ['name' => 'C Cocktail'],
+        ]);
+
+        $response = $this->getJson('/api/cocktails?sort=name');
+        $response->assertJsonPath('data.0.name', 'A Cocktail');
+        $response->assertJsonPath('data.1.name', 'B Cocktail');
+        $response->assertJsonPath('data.2.name', 'C Cocktail');
+
+        $response = $this->getJson('/api/cocktails?sort=-name');
+        $response->assertJsonPath('data.0.name', 'C Cocktail');
+        $response->assertJsonPath('data.1.name', 'B Cocktail');
+        $response->assertJsonPath('data.2.name', 'A Cocktail');
     }
 
     public function test_cocktail_show_response()
@@ -82,7 +154,7 @@ class CocktailControllerTest extends TestCase
                 ->has('data.short_ingredients', 3)
                 ->has('data.ingredients', 3, function (AssertableJson $jsonIng) {
                     $jsonIng
-                        ->has('id')
+                        ->has('ingredient_id')
                         ->where('amount', 60)
                         ->where('units', 'ml')
                         ->where('optional', false)
@@ -156,7 +228,7 @@ class CocktailControllerTest extends TestCase
                 ->where('data.garnish', 'Lemon peel')
                 ->has('data.ingredients', 2, function (AssertableJson $jsonIng) {
                     $jsonIng
-                        ->has('id')
+                        ->has('ingredient_id')
                         ->etc();
                 })
                 ->etc()
@@ -216,15 +288,6 @@ class CocktailControllerTest extends TestCase
     public function test_user_shelf_cocktails_response()
     {
         $response = $this->getJson('/api/cocktails/user-shelf');
-
-        $response->assertStatus(200);
-
-        $response->assertValidResponse(200);
-    }
-
-    public function test_user_favorites_cocktails_response()
-    {
-        $response = $this->getJson('/api/cocktails/user-favorites');
 
         $response->assertStatus(200);